Revised data type of id in user; Added callback on open of database;

+++ app/src/main/java/com/fujiyama/pulp/sqlitedb/configs/sqlite/schema/tables/User.java
@@ -5,25 +5,23 @@ import android.arch.persistence.room.Entity;
 import android.arch.persistence.room.PrimaryKey;
 import android.support.annotation.NonNull;
 
-import java.math.BigInteger;
-
 @Entity(tableName = "USERS")
 public class User {
 
     @PrimaryKey(autoGenerate = true)
     @ColumnInfo(name = "ID")
     @NonNull
-    private BigInteger mId;
+    private int mId;
 
     @ColumnInfo(name = "FULL_NAME")
     private String mFullName;
 
-    public User(BigInteger mId, String mFullName) {
+    public User(int mId, String mFullName) {
         this.mId = mId;
         this.mFullName = mFullName;
     }
 
-    public BigInteger getmId() {
+    public int getmId() {
         return mId;
     }
